requirements.txt
unsloth[torch]
torch
datasets
transformers
accelerate
bitsandbytes
flash-attn
pyyaml

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1